+++ src/Module/ModuleUtil.php
@@ -8,6 +8,7 @@
 
 namespace HeimrichHannot\UtilsBundle\Module;
 
+use Contao\CoreBundle\Framework\ContaoFrameworkInterface;
 use Contao\Module;
 use Contao\ModuleModel;
 use Contao\System;
@@ -15,6 +16,17 @@ use Model\Collection;
 
 class ModuleUtil
 {
+    /**
+     * @var ContaoFrameworkInterface
+     */
+    private $framework;
+
+    public function __construct(ContaoFrameworkInterface $framework)
+    {
+
+        $this->framework = $framework;
+    }
+
     /**
      * Get the class name of a given module.
      *
@@ -45,23 +57,37 @@ class ModuleUtil
      */
     public function isSubModuleOf($module1, $module2): bool
     {
-        if (!\is_string($module1) || false === strpos($module1, '\\')) {
-            $module1 = $this->getClassByModule($module1);
-        }
-
-        if (!$module1 || !class_exists($module1)) {
+        $module1 = $this->getModuleClass($module1);
+        if (empty($module1)) {
             return false;
         }
-
-        if (\is_string($module2) || false === strpos($module2, '\\')) {
-            $module2 = $this->getClassByModule($module2);
-        }
-
-        if (!$module2 || !class_exists($module2)) {
+        $module2 = $this->getModuleClass($module2);
+        if (empty($module2)) {
             return false;
         }
 
         return is_subclass_of($module1, $module2);
+    }
+
+    /**
+     * Get the full qualified class name for a given module.
+     *
+     * @param ModuleModel|Module|string $module A module object, a module model object, a full qualified model class name or model type.
+     *
+     * @return string
+     */
+    public function getModuleClass($module)
+    {
+        if ((is_string($module) && !class_exists($module)) || $module instanceof ModuleModel) {
+            $module = $this->getClassByModule($module);
+        }
+        if (is_object($module)) {
+            $module = get_class($module);
+        }
+        if (!$module || !class_exists($module)) {
+            return '';
+        }
+        return $module;
     }
 
     public function getModulesByType(string $type, array $options = []): Collection